Corp 46 delete carpool (#22)
* Corp-22-get_availableRides

Added getAllRides method in Mariadb service.

* Corp-22-get_availableRides

Added getAllcarPools method in CarpoolService.

* Corp-47-Join_and_approve_carpooling

Added Join and approve carpooling method in maria db service and added jain carpooling method in booking service

* Corp-47-Join_and_approve_carpooling

Added approve passenger method in bookingService and refactored code in some classes

* Corp-46-Join_delete_carpool

Added deletecarPool method in mariadb service and in CarPoolService.

* Corp-46-Join_delete_carpool

Added  validation for deletecarPool method in CarPoolService.

* Corp-46-Join_delete_carpool

solved merge conflicts

+++ bookingservice/src/main/java/com/example/bookingservice/service/TripService.java
@@ -44,7 +44,7 @@ public class TripService {
         }
 
         Carpool carpool = getCarpoolById(passengerOptional.get().getCarpoolId());
-        if(carpool.getDriverId() != driverId) {
+        if(carpool.getDriverId().longValue() != driverId.longValue()) {
             throw new RuntimeException("Not your Passenger");
         }
